// src/API.ts
export type Post = {
	__typename: 'Post';
	id: string;
	title: string;
	content: string;
	username?: string | null;
	coverImage?: string | null;
	comments?: ModelCommentConnection | null;
	createdAt: string;
	updatedAt: string;
};

export type ModelCommentConnection = {
	__typename: 'ModelCommentConnection';
	items: Array<Comment | null>;
	nextToken?: string | null;
};

export type Comment = {
	__typename: 'Comment';
	id: string;
	message: string;
	post?: Post | null;
	postID: string;
	createdAt: string;
	updatedAt: string;
	createdBy?: string | null;
};